More cosmetics.
KML elements are in alphabetical order now, I believe.
